added features

+++ src/main/java/ru/geekbrains/spring/market/model/dtos/CartDto.java
@@ -2,8 +2,9 @@ package ru.geekbrains.spring.market.model.dtos;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.geekbrains.spring.market.model.entities.Cart;
+import ru.geekbrains.spring.market.beans.Cart;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -11,7 +12,7 @@ import java.util.stream.Collectors;
 @Data
 public class CartDto {
     private List<OrderItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
     public CartDto(Cart cart) {
         this.totalPrice = cart.getTotalPrice();
