+++ src/main/java/ru/geekbrains/spring/market/model/dtos/OrderItemDto.java
@@ -4,13 +4,15 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.geekbrains.spring.market.model.entities.OrderItem;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 @Data
 public class OrderItemDto {
     private String productTitle;
     private int quantity;
-    private int pricePerProduct;
-    private int price;
+    private BigDecimal pricePerProduct;
+    private BigDecimal price;
 
     public OrderItemDto(OrderItem orderItem) {
         this.productTitle = orderItem.getProduct().getTitle();
